src: Introduce guard module with PreemptGuard structure
This contrasts with the continuation module's old implementation: now, allowing a guard to go out of
scope automatically unblocks the signals again, but ONLY IF THEY WERE BLOCKED TO BEGIN WITH.  (This
is desirable because when interposing on library calls, we won't want to accidentally enable
preemption on a thread where it wasn't before!)  To suppress the check, explicitly invoke
PreemptGuard::unblock() where you would have allowed your instance to be drop()'d.  To block without
a corresponding unblock, immediately forget() the (stack-allocated) result of PreemptGuard::block().

Note that despite having been switched to the new implementation, CallStacks continue to follow the
old behavior of more or less unconditionally unblocking (unless we're in a signal handler or panic).
This is actually desirable because they should only ever be used from threads that *have* a
CallStack (and further, one that is in active use), and such threads are running preemptible code.