src/ucontext.rs: Replace getcontext() with a new API that returns a context
instead of accepting a mutable reference to it.  On subsequent returns through the code path, it
instead returns None to indicate that this is not our first visit.  The main advantage of this new
approach is that it obviates the need for extreme care when handling the context's (de)allocation:
instead, the caller can save any returned context wherever it desires, simply skipping this step if
no new context was just created.