src: Swap contexts' fpregs sections in-place rather than assuming their pointers should be adjusted
It appears that signal handlers (and sigreturn()) expect the fpregs to be located in a slightly
different location that overlaps beyond the end of the ucontext_t argument.  This patch seems to
eliminate the segfaults associated with violating this assumption.

Note that this change makes a new assumption that the fpregs pointers in the ucontext_t s are valid
going into the swap(); as such, it also amends UntypedContinuation::new()'s behavior to
fixupcontext() as it boxes the argument.  We leave the fixupcontext() in setcontext() because
memmove()s on return from getcontext() and makecontext() would still cause problems otherwise.