src: Use a trampoline function to save the context when preempting rather than storing the signal handler's
because it appears POSIX may have withdrawn the guarantee that setcontext() works with handlers'
arguments before it removed the rest of ucontext.h from the standard.

Caveat emptor: this approach currently overwrites part of the stack below where the signal handler
is running, and may thereby corrupt it.  If we stick with it long-term, we'll want to use
sigaltstack().