src/ucontext.rs: Fix a double free bug that was causing release builds of the tests to crash
On the second path through the checkpoint() routine (i.e. after client code had restored the
context), the Rust compiler didn't see any use of the context object, and so attempted to drop it
when it went out of scope.  However, before the time travel, ownership of this object had been
passed to the first closure, which had already subsequently deallocated it via setcontext()!

We now pass the tests for both types of builds!