Make setcontext() take a shared reference to the Context instead of acquiring ownershipv of it
This is possible because of the platform setcontext()'s guarantee that it will not modify the
context (and therefore that said context may be restored multiple times).  (Note that unfortunately
POSIX defined setcontext() with a const-incorrect signature.)  This implementation requires interior
mutability so that setcontext() is still able to perform the MoveInvariant fixup.

Note that the change causes UB in at least the killswap_sigsetcontext() test, which now segfaults.