Detect illegal intra-stack use of nested contexts by using a per-thread shadow stack of guards
This change allows us to invalidate any nested getcontext() invocations' contexts when we return
from a call to getcontext() or when we make an intra-stack call to setcontext().

This should also pave the way to leak-free cleanup after setcontext(), which is a concern now that
this function accepts a reference rather than a value.