Attempt to store successor context references in contexts returned from makecontext()
The goal behind this change was to alleviate the makecontext() leaks by enabling setcontext() to
follow the resulting linked list, validating and destructiong (if on an orphaned stack) the
contexts in the chain of "succession."  Unfortunately, this approach doesn't work because the
compiler knows that UnsafeCells, like mut references, are invariant in their type; as such, it
becomes impossible to call makecontext() on a valid checkpoint from getcontext() (i.e. from within
the closure passed thereto).