handle.c: Fix size calculation when unprotecting immutable GOT entries
This resolves a bug where later writes to .data/.bss resulted in a protection fault.

It's important to include only .got pages but not .got.plt pages, since the latter is required to be
mutable and therefore might share pages with other mutable sections!  We also add an assertion to
verify that the division between the two sections does, in fact, lie on a page boundary.