handle.c: Always populate the base namespace's GOT's link_map during handle initialization
This ensures we always have access to it from load_shadow(), which will be necessary if we want to
do more complex relocation entry processing in the base namespace.  (Before, it could've been absent
in the case of BIND_NOW-style eager relocation processing, or in the case of object files that had
GLOB_DAT relocation(s) but no JUMP_SLOT ones.