Add another benchmark to compare the overhead of library multiplexing with memoization
Disabling memoization requires placing the resolved address at a different location depending on
whether we're using multiplexing: without, it only needs to reach the Rust call site, whereas with,
it needs to be placed within the shadow GOT so it can be used by the PLOT trampolines.  Because
it's important to restore the latter entry as soon as we're done, we redesigned both of the "lazy"
(i.e., memoization-free) benchmarks to use a single with_eager_nop() wrapper rather than exposing
the nop_location() function of yesteryear.