Build system support for creating a source archive of an arbitrary revision of the tree
This makes it easy to test or benchmark without uncommitted changes, as well as to regression-test
past snapshots (by specifying a custom REVISION, such as @{u}) when a full bisect isn't warranted.
The generated archive includes a stub "Git repository" to fool the clean target into working as
expected.  It also contains a VERSION file with a record of its corresponding revision.