Skip installing trampolines over GOT entries that already contain NULL
This solves the problem of "client" code that NULL-checks GLOB_DAT relocations before invoking them
in the general case: the NULL-check remains accurate, so they will no longer attempt to invoke
trampolines that bridge to nonwhere.  Thus, we are also able to remove the preload quirks for libc
initialization and valgrind teardown.