Support GLOB_DAT relocations pointing to *data* objects defined by the same object file
Previously, shared objects that included an overrideable static were handled incorrectly: because
such statics are internally referenced via GLOB_DAT relocations, we would replace their GOT entries
with the address of an inapplicable PLOT trampoline function.

We now consult an object file's section table and decide whether to install a trampoline depending
on whether the symbol's section is marked executable.

Note that this does not (yet) handle the case where a shared object accesses a static data object
defined by the *executable*, in which case the shared object will contain a GLOB_DAT to an
*undefined* symbol!