Support allocating multiple PLOTs/GOOTs per object file so we can handle arbitrarily-large libraries
This changes plot_insert_lib() to generate an array of PLOT pointers.  If this list contains
multiple elements, all preceding ones are filled to capacity exclusively with elements from the
object file in question, and have special adjustment fields in their GOOTs set to the offset that
should be added to each of their index values.  The last element, which is identifiable by its
GOOT's identifier that matches the handle's shadow.override_table, uses the handle's
shadow.last_adjustment as its offset, which must then be corrected *downward* by the
value of shadow.first_element.