handle.c: Fix design mistake where GLOB_DAT relocations hit unpopulated SGOT entries during setup
This caused some of our own library calls to be rerouted to NULL() instead.  In order to avoid this,
we preallocate and prepopulate the first portion of the base namespace's shadow GOTs ahead of time
in handle_init() so they are all ready by the time handle_got_shadow() is first called.