handle.c: Fix unsafety of calls into the vdso and dynamic linker from ancillary namespaces
Because we were completely stubbing out the vdso and dynamic linker handles, there were no
trampolines and calls into them continued unintercepted.  In the case of the dynamic linker, this
was unacceptable because it meant we never recorded a namespace switch back to base (which happened
automagically because the dynamic linker maintains itself as a special singleton).

Instead of creating empty dummy handles for these two special object files, we now partially
populate handles for them, generate trampolines, and forge usable shadow GOTs.