Support for intercepting accesses to globals defined in other object files
This slow path is an alternative to intercepting dynamic function calls that is necessary to achieve
consistent semantics.  It works by shadowing data symbols with inaccessible memory addresses so that
client code will generate a segmentation fault upon performing an access.  We catch this signal,
disassemble the instruction (using the LGPL libasm from elfutils) to find the register that
contained the base of the address calculation, and attempt to look up its shadow GOT entry for the
current namespace (which is possible because SIGSEGVs in particular are guaranteed to be delivered
to the causing thread).  If we're successful, we replace that register's contents with the shadowed
address so the access will succeed when retried.  Otherwise, we defer to whatever signal handler was
previously installed.  (Note that we don't yet properly reroute callers' attempts to change the
handler, so they can later defeat this scheme at the moment.)

Although this adds support for this feature, it does not yet enable it: handle_init() doesn't
actually add STT_OBJECT symbols to the "trampoline"-creation todo list.