globals.c: Add heuristic to handle situations where client code has recently precomputed a pointer
This handles the common cases where client code stores a pointer to a location internal to a global
defined in a different object file.  Such cases are challenging because the base register of the
address calculation performed by the faulting instruction doesn't contain a real multiplexing
address.  In such situations, we apply a heuristic: if the code faulted at a shadowed address
"recently" (as judged by its register contents still being unchanged), we retrieve the address
it resolved to, treating the newly-faulting address as an offset into that structure.

This resolves a crash we were seeing within the getenv() call from the assertion when handling
preloaded object files during initial shadowing.  The mirror test now passes with shadowing of
globals enabled.