handle.c: Do not shadow read-only globals
These variables don't change during the program's normal lifetime, so we assume they're also
consistent between copies of the library.  This alleviates problems such as attempted infinite SEGV
assertion when sysmalloc() is enlarging the heap and the disassembler needs to allocate, too.

With this change, nginx runs again, even with shadowing of globals enabled!