Prevent references to functions from ancillary namespaces from creeping into the dynamic linker via constructors
This was inspired by dynamic linker modifications performed by libpthread, even under dlmopen().
Left unaddressed, this issue at least hurts debugabbility, if not correctness.  It's tough to
reason about, but it feels like potentially hazardous behavior.

This also introduces a warning if any affected objects are not whitelisted, which is also handy
because it catches scary situations where the interpreter is not captured by the whitelist.  We add
libdl.so back to the whitelist to squelch this warning.