Restructure build system to perform an intermediate Rust link into a consolidated object file
This approach allows us to demote most of the global symbols to locals before performing the final
link, giving us finer-grained control over what gets exported.  Until the churn settles, we're
temporarily dropping support for building final targets other than the preload library.

Eventually we'll have an additional Rust crate providing the public API and managing the final link.
In the meantime, we're keeping the gotchapreload.c stopgap around, with the caveat that it currently
exports a few symbols that should be private.