Interpose attempts to *mask* SIGSEGV as well
This adds handling for the signal() and sigprocmask() library calls as well.  Note that the use of
fputs()/stderr required rebuilding the libgotcha_repl module as position-independent, which
mysteriously also stopped GCC from outputing any relocations for the "recursive" calls up to the
shadowed implementations.  Adding a weak aliase for each interposed symbol fixes this, causing both
GCC and Clang to generate relocations that the linker targets at the PLT.  We then convert these to
local symbols in the build system so they won't be exported.  One nice side-effect of this symbol
requirement is that debuggers can use them to unambiguously break on our implementation.