handle.c: Fix bug where we applied forced interpositioning to ourselves if we were an ordinary program dependency
This resulted in our own library calls erroneously being redirected to our interposed wrappers,
which then recursed infinitely.  Because handle_got_shadow_init() modifies outgoing calls but
handle_got_whitelist_all() does not, the bug only afflicted the former; this meant that the
problem occurred when we were a (transitive) dependency of the main executable but not when we (or
a dependent) were LD_PRELOAD'd.