gotcha.mk: Fix horrible static library double-inclusion bug
Due to an accidental use of $^ instead of $<, our static library was appearing twice in the
dependency list of client libraries: once normally, and once with --whole-archive.  Only the latter
had any effect unless the client library referenced any of our symbols, in which case it pulled in
duplicates of a subset of our definitions.  This resulted in linker errors that, if manually
overridden,
could lead to a client library with multiple copies of our constructor.