handle.c: Fix a bug where intra-library GLOB_DAT calls to individually-whitelisted functions were directed to the wrong function altogether
This happened because, as an "optimization," we skipped processing such relocations in the base
namespace, and therefore didn't create an entry in the temporary globdats lookup table.  This patch
addresses preserves the optimization but addresses the problem by only skipping the GOT update at
the end of processing in such a case.

The problem was discovered by calling asprintf() in an ancillary namespace, since that makes a
libc-internal dynamic call to malloc().