Remove interposition modules: stdlib (dynamic allocator) and pthread (thread creation)
This fixes an outstanding segfault when running a debug build of the test harness.  The module is no
longer correct or necessary: most of its code exists to deal with the fact that the interposers call
dlsym() to find the functions they wrap, but this function may call them back, resulting in infinite
mutual recursion.  The next generation of the module won't need to call dlsym(), though, thanks to
libgotcha's support for pure symbol-based interposition.  Furthermore, we won't need to interpose
these functions to disable preemption since libgotcha will handle that automagically on library
group transition; instead, we'll only need to perform allocation tracking once we support resource
cleanup.

The *unit* test driver refuses to run due to the libgotcha linkage; skip it using:

	$ cargo build --test inger

Note that the unbundled test crate still appears to segfault when built in release mode.