handle.c: Skip running suspected monkey-patching legacy constructors in ancillary namespaces
This resolves a deadlock within the dynamic linker when attempting to reinitialize a namespace that
includes libpthread.so.0.

It obviates the need to call the base namespace's libpthread constructor again after loading
ancillary copies.  However, it may leave the ancillary copies inoperable, and therefore probably
only works because we whitelist almost every call to libpthread.