src/linger.rs: Rewrite preemptible functions (again) to migrate more code to non-generic functions
This minimizes the amount of code that is subject to unexpected dynamic calls interposed by
libgotcha.  Morally, it implements this workaround with relative elegance by essentially outsourcing
more work to be executed on the oneshot execution stack, which must be type-blind anyway.  (In
reality, though, there are a lot of helper functions, too.)