Make the trampolining lock a per-namespace attribute rather than a per-thread one
This fixes a severe bug where the shared-code callback could be skipped when multiple preemptible
functions were in flight on the same kernel thread:  If one such function was paused within its
shared-code callback function, that function wouldn't be called for *other* preemptible functions
on that thread until the first was resumed and allowed to run to a different point.  If the first
function was instead canceled, its thread would *never* invoke the callback again!

In addition to reassigning ownership of the trampolining lock, this patch explicitly releases it
during cancellation as part of the namespace reinitialization process.