src: Add a simple profiler, since it's sometimes hard to get full-featured ones working with libgotcha
To use it, use with_profile() to call begin() and end() from two distinct points in your program.
Assuming the call sites form at least one in-order pairing, the average latency between them will be
printed out when the program exits.  The only non-obvious thing is that, with LIBGOTCHA_NOGLOBALS=,
it only works if the PLT entry of the call to start() first resolves from outside any ancillary
libset; to work around this issue, just call it once from outside a timed function if you otherwise
only use it from inside it.