init.c: Fix constructor's bootstrapping protection to avoid false postives/negatives especially under debuggers
The old heuristic would sometimes fail when running under valgrind or rr, resulting in long,
frustrating debugging sessions where a wild goose chase would reveal that multiple copies of
libgotcha were competing to break everything.

It may be helpful to export this function in our API for client libraries to leverage for avoiding
the same issue.