src/linger.rs: Make Linger and Continuation types tolerate dynamically-sized closure types
This will allow us to use trait objects to store type-erased closures, which will be necessary to
implement the C language foreign-function interface without depending on generics.

It was surprisingly simple to make this change without making Continuation a DST because the
launch() implementation already heap allocated the functional member; this was required in order
for its address to be stable because we might move the Continuation instance between preempting and
resuming it.