src/linger.rs: Support erase()'ing the generic types from a Linger and resume()'ing it afterward
Note that launch() still never hands out an erased Linger directly: while that approach seems like a
more obvious one, there's no way to instruct it to adopt this behavior only some of the time because
return-position impl Trait is not influenced by external type inference, and because nesting the
closure two levels deep within the Continuation prohibits coercion to a (fat) trait object pointer.

The MaybeUninit business is necessary to move out of Continuation because it implements Drop.