Make our own calls to __tls_get_addr() aware of our own custom calls to arch_prctl()
This introduces a static interposition module whose operation is described in its header.
Currently, it is used to make thread-local variables in each of our *own* modules that is compiled
*without* the initial-exec TLS model ignore any manual replacement of the TCB that *we* perform.
(Furthermore, it provides special-case handling to copy the thread's namespace setting, which cannot
use __tls_get_addr() for fear of clobbering registers during dynamic function dispatch.)

Specifically, for each custom TCB we track its corresponding "parent" TCB; this is always a real
kernel thread's TCB, regardless of how many times the TCB has been replaced in the meantime.  As a
performance optimization, we also cache a self-reference to the current TCB to save an extra syscall
per __tls_get_addr() call.  (If this approach later proves error prone, we could instead read it
back from the segment register at the cost of additional scaffolding.)

There are presently no plans for us to actually modify the TCB; rather, this is intended as a step
toward allowing client libraries to do so without breaking our runtime in the process.