src/ucontext.rs: Add an assertion to catch pthread_kill() "successes" that don't deliver a signal
This makes it obvious when pthread_kill() doesn't do anything, which might happen because the
SIGSETCONTEXT signal is masked for the current thread, or because something is wrong with libpthread
and/or the current thread's identifier.

Such a situation was previously very difficult to debug because it resulted in a Success error.