benches/bench.c: Debugging benchmark program used to demonstrate an accumulating slowdown
It turns out that TLS isolation introduced a bug causing libinger to create a new POSIX timer on
every launch(), rather than every launch() on a new thread!  Checking this reduced benchmark in
since it helped reveal this bug where the ordinary benchmark wouldn't converge.  It also integrates
with the profiler from Google's gperftools, although that turned out not to be useful here since the
performance degradation was in system time.