README.md: Update recommended glibc build configuration
We have found a static surplus multiplier of 2000 to be sufficient except when leaning on the
current (leaky) cancel() implementation.  It's important not to set it too high, as this slows
down TCB allocation and thread creation.  Also note that both operations also scale with the number
of available libsets; experimental mitigations such as block-allocating TLS areas appear to win over
the default glibc approach by the time you have 125-150+ libsets.