libgotcha.gdb: Debugging script for automatically loading symbols from ancillary namespaces
This fixes the debugger's ability to display stack traces, source code, etc. when stopped outside
the main linker namespace.  It works as long as you are not in an ancillary copy of the *executable*
itself (as GDB does not support loading additional symbols from an executable, so we skip it).

In order to use this script, you must add the dlfcn/ and elf/ directories from the glibc source tree
to your search path via GDB's dir command; for instance:
	$ gdb -x libgotcha.gdb -ex dir\ .../path/to/glibc/dlfcn:.../path/to/glibc/elf

Tested on the following minimal test cases with gdb 9.1, rr 5.4.0, and valgrind 3.16.1...

C:
	#include "libgotcha_api.h"

	#include <stdlib.h>

	int main(void) {
		libgotcha_group_thread_set(libgotcha_group_new());
		abort();
	}

Rust:
	extern crate gotcha;

	use gotcha::Group;
	use gotcha::group_thread_set;
	use std::process::abort;

	fn main() -> Result<(), ()> {
		group_thread_set!(Group::new().ok_or(())?);
		abort();
	}