testsuite: Remove vestigial signal() intercepter stub
This was added as a workaround in an effort to get glibc's own test suite to run, which revealed
that we currently break fork()'s child handler (registered via the internal interface backing
pthread_atfork()).  In the end we went with gnulib for now because it automagically skips tests that
won't pass because of the way the installation is configured, and seems muck less flaky in general.