Move arch_prctl() logic into tcb module to permit its use from our custom API as well
Currently, we cannot know which namespace is making a call to arch_prctl(), since the call gets
trampolined into namespace 0 before arriving in our implementation; this means we must assume the
call came from outside namespace 0 and reinitialize all ancillary namespaces' TCB regions.  This is
expensive, and we'd like to expose a custom interface that allows the caller to declare which
ancillary namespace, if any, is actually in use.