When handles_restoretls() encounters an unaddressable TLS region, force the allocation of its DTV
We used to fail the assertion on modules with static TLS relocations (e.g., ourselves, libc,
relevant executables) because the dlinfo() interface relies on DTV entries to find the base address
of the TLS range.  For such modules, the dynamic linker lazily initializes the DTV since it is only
needed for dynamic TLS access models.  By explicitly calling __tls_get_addr() on the first TLS
symbol, we force said initialization to happen immediately.