testsuite/testinger.c: Remove getenv() bootstrapping interposition
This was initially added to fix a bootstrapping problem whereby libinger's constructor was observed
calling getenv() before libc had initialized its metadata, apparently because libtestinger was being
preloaded; however, this issue no longer appears to be reproducible.

Because the interposition was not removed after bootstrapping was complete, it was breaking
environment variable–related tests and is being removed.