plot.S: Refactor to combine the split hook_trampoline() and hook_ret() procedures
Previously, these were unstructured routines with a jump in between them.  Furthermore, the fact
that they received "arguments" on the stack made it difficult to reason about alignment.  To fix
both these shortcomings, we introduce a hook_retpoline() function whose sole purpose is to serve as
an extra return address on the stack, thereby correcting the alignment for both the whitelisted
library function and any installed callback hook.