plot.S: Stop clearing the caller namespace on trampoline return
This will be needed to support interruptible subregions: otherwise, the inner trampoline would trick
the outer one into re-"restoring" the namespace to 0.  Leaving a leftover value in caller, as this
patch does, is no big deal because it will be cleared the next time anyone on that thread calls a
shared function.  Since such a shared call is the next time we could possibly be running, we will
never read the garbage value before it is clobbered by procedure_linkage_override().