Port to Rust 1.37
As of this version, the Rust compiler started passing a version script when compiling dynamic
library scripts in order to avoid exporting non-public symbols.  Unfortunately, this change broke
our dynamic interposition functions in the libgotcha_repl module because rustc was not aware of
their symbol names.  The absence of these symbols altered our semantics and caused nontrivial
control libraries to crash.  It was detectable a priori by objdump'ing the dynamic symbol table of
an affected shared object file and observing that there were no entries for our interpositions.

Informing Rust of each replacement function would require maintaining a parallel list that could get
out of date and lead to subtle bugs.  Instead, this patch introduces a wrapper script, cc, that
passes through all command-line switches except for -Wl,--version-script.  It's named 'cc' as a
reminder that the filename is significant to rustc (e.g., using 'ld' will cause the compiler to pass
different switches).  The client library Make and Cargo integrations are also updated; in the latter
case, an intermediate stub 'linker' script is used in case particular client libraries want to
interpose their own custom script in front of ours (e.g., to pass a custom interpreter).