Fix deadlock in Rust 1.50.0 and later
This version of the standard library introduces a "weak" syscall that tries dlsym() while
bootstrapping the std::sys::unix::rand module.  Unfortunately, this codepath called from HashMap's
constructor, so our dlsym() interposition gets called from within trampolines()' Once closure; the
interposition then calls back into trampolines() to search for a PLOT corresponding to the found
symbol.

To avoid this mutual recursion, this patch introduces a conditional-compilation configuration string
to indicate whether the once_is_completed feature---introduced in Rust 1.43.0---is available, and
locks out the PLOT search during bootstrapping if so.  The feature detection is based on an
extensible shell script that might support more features (or configuration strings, for that matter)
in the future as necessary.