handle.c: Port to glibc 2.30
In this version, ld.so stopped accepting position-independent executables in dlopen() because it's
impossible to correctly handle their COPY relocations or call their constructors.  We don't support
COPY relocations anyway, and the constructor issue is comparatively insignificant, so this patch
applies the same shared object editing hack we use for NODELETE libraries.