handle.c: Fix libgotcha_group_symbol() breakage on glibc 2.30
This alters the previous patch's behavior so that it leaves our executable's h->path alone so
subsequent uses of handle_progname() are not dangerous.  This was causing libgotcha_group_symbol()
to obtain a NULL link map when requesting ancillary namespace symbols, for instance.

Note that we are still seeing separate libgotcha_group_symbol() breakage on some new ld.so versions
(e.g., 2.33) when global variable interception is *enabled*: it seems that dlsym() passes a garbage
but *readable* pointer to __rtld_lock_lock_recursive().