globals.c: Extract the offset component of the displacement-mode address alongside the base register
This is not currently used for anything, but may be needed to find the PLOT page in cases where the
symbol being dereferenced falls off the edge of the single corresponding unreadable trap page that
we currently allocated.  Whether such accesses are currently intercepted at all is currently
unspecified because their "bogus" placeholder addresses may actually be readable!

This problem occurs reproducibly when dlsym() attempts to construct a pointer to ld.so's load
biglock in recent glibc builds, likely due to a tweak to the _rtld_global structure in glibc 2.31
that appears to make its size scale with DL_NNS, such that the symbol is commonly much larger than
a page.  Actually fixing this will require allocating additional unreadable trap pages.