plot.h: Fix libgotcha_group_symbol() breakage on glibc 2.31 with global variable interception enabled
This resolves the supposedly-unreadable pointer that had flowed off the end of the fully-protected
PLOT footer page by simply mapping two footer pages instead.  Of course, the robust way to fix this
would be to compute the number of trailer pages required for each library (ideally as a function of
its placement in the GOOT) by finding the maximum size of its STT_OBJECT symbols...

However, this simple fix was successfully tested by making a libgotcha_group_symbol() call against
namespace 1 atop a glibc 2.33 build with DL_NNS set to 512.  Given that the _rtld_global symbol is
about a page, it is probably one of the largest symbols commonly encountered in modules' ABIs.