Actually fix rr dlsym() bootstrapping crisis
This patch fixes rr 5.4.0, which is afflicted by the bug described in the preceding one.  It does so
by making direct calls to:
 * __libc_dlopen_mode(): to avoid dlopen(), which we are resolving
 * __libc_dlclose(): to fix the refcount, à la namespace_get()
 * __libc_dlsym(): to avoid dlsym(), as dynamic interpositions are not yet initialized

Besides fixing the crash in the dynamic linker caused by the previous attempt, this approach has the
advantage of also working on eagerly-resolved dynamic executables (i.e., those marked with DF_NOW).

We once again support rr, so long as $LIBGOTCHA_NODYNAMIC is defined!