Add experimental libgotcha_shared_prehook() mechanism for running a callback on shared code *entry*
This complements the existing callback mechanism for shared code *exit.*  For now, it is implemented
by installing a signal handler to avoid having to manually preserve all the registers.  This works
because the signal is thread-directed as per signal(7), but imposes the restriction that the
callback may only use async-signal-safe library routines.