testsuite/build: Work around issue where running cargo build twice rebuilds from scratch
This appears to stem from an update to one of the build tools we depend on.  While normally
harmless, it caused the testsuite build script to complain about "multiple definition of `signal,'"
"undefined reference to `libgotcha_dlsm,'" and "undefined reference to `libgotcha_signal'" because
cargo was overwriting the libgotcha rlib file after we objcopy'd it.

If we ran into a similar problem in the future, another approach would be to back up the libgotcha
rlib after objcopying and restore it after the failed build.