compiler: Add scaffolding for writing LLVM llc codegen passes out of tree
This assumes that you have the following installed:
 * the LLVM sources, via your OS's package manager
 * the LLVM tools, via the upstream Rust installer

Then you can build and run the pass as follows:
 $ echo "fn main() {}" >true.rs
 $ rustc --emit llvm-ir true.rs
 $ make
 $ LD_PRELOAD=./libingerc.so .../llc true.ll

Note that if you want to run llc in a debugger, you probably want to build it with debug symbols.
In this case, you must be sure to tell the LLVM build system to build it dynamically linked.  You
should also disable the ABI breaking checks; however, if you forget to do this when building LLVM,
you can override it by building the plugin like so:
 $ make CPPFLAGS=-DLLVM_DISABLE_ABI_BREAKING_CHECKS_ENFORCING