compiler: Support generating an X86RegisterInfo.h header for exposing x86-64 register enums
This information is about to become important to our pass.  Unfortunately, it is stored in the
X86.td table in the LLVM source tree, which is not distributed as part of all distros' source
packages.  It is, however, available in the Rust source distribution.  As a consequence, if the file
needs to be regenerated, one may need to install the upstream Rust source package and do:

	$ make LLVM_SOURCE_DIR=.../lib/rustlib/src/rust/src/llvm-project/llvm