Fix arch_prctl()-breaking regression caused by moving the parent namespace from the stack to the TCB
The simplification made in 3d5806f was indeed valid, but only because at that time the caller's
namespace was stored on the stack and therefore unaffected by the TCB switch.  Unfortunately, this
design assumption later changed in 28c4eb2 when we moved this information to the TCB to provide an
tcb_parent() function to support the upcoming libgotcha_group_caller() API.

This change also had to undo the very minor patch a7fa062.

Here's a minimal example that confirmed the design mistake and breakage:

	pthread_t newtcb;
	pthread_create(&newtcb, NULL, halt, NULL);

	libgotcha_group_t libset = libgotcha_group_new();
	libgotcha_group_thread_set(libset);

	arch_prctl(ARCH_SET_FS, newtcb);
	assert(libgotcha_group_thread_get() == libset);